requirements.txt
pageindex==0.2.8
openai==2.30.0
python-dotenv==1.2.2
pydantic==2.12.5
requests==2.33.1
httpx==0.28.1
anyio==4.13.0
sniffio==1.3.1
